third_party/blink/web_tests/platform/generic/fast/forms/select/popup-with-display-none-optgroup-expected.txt

Option Groups with display: none should not show the children option elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS select[1].value is "Four"
PASS select[1].value is "Six"
PASS select[0].value is "Six"
PASS successfullyParsed is true

TEST COMPLETE

One
Two
Three
Four
Five
Six
Seven
 
One
Two
Three
Four
Five
Six
Seven
